refactor #19: 타입을 모아서 관리
