+++ src/component/main/main/list/pagenation.tsx
@@ -1,6 +1,6 @@
-import type { PaginationState } from '../../../../types/main/main';
+import type { PaginationProps } from '../../../../types/main';
 
-const Pagenation = ({ page, totalCount, setPage }: PaginationState) => {
+const Pagenation = ({ page, totalCount, setPage }: PaginationProps) => {
   return (
     <div className="fixed bottom-0 flex justify-center gap-2 rounded-full bg-white/90 px-4 py-2 backdrop-blur">
       <button
